Move file under proper folder
